@@ -8,9 +8,9 @@ dependencies:
   flutter:
     sdk: flutter
 
-  http: ^0.12.0+2
-  shared_preferences: ^0.5.1+2
-  flutter_webview_plugin: ^0.3.3
+  http: ^0.13.3
+  shared_preferences: ^2.0.7
+  flutter_webview_plugin: ^0.4.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
@@ -29,7 +29,7 @@ flutter:
           weight: 700
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
   # To add assets to your package, add an assets section, like this:
   # assets:
