@@ -10,16 +10,14 @@ class GenericSlackButton extends StatelessWidget {
 
   final String clientId;
   final String clientSecret;
-  final String redirectUrl;
 
   const GenericSlackButton(
-      {@required this.clientId,
-      @required this.clientSecret,
-      @required this.onSuccess,
-      @required this.onCancelledByUser,
-      @required this.onFailure,
-      @required this.onTap,
-      this.redirectUrl});
+      {required this.clientId,
+      required this.clientSecret,
+      required this.onSuccess,
+      required this.onCancelledByUser,
+      required this.onFailure,
+      required this.onTap});
 
   bool get enabled => onSuccess != null;
 
